feat: Modify domain entities.
Subsitute List for Set. Reorder elements is lost though uniquness on productList is gained. Id type changed to Long

+++ src/main/java/com/ecommerce/domain/Article.java
@@ -10,6 +10,12 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
+/**
+ * Represents a phyisical Article
+ * @author Jose Antonio Lopez
+ * @version 1.0
+*/
+
 @Entity
 @NoArgsConstructor
 @Getter
@@ -19,7 +25,7 @@ import lombok.ToString;
 public class Article {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 	private String description;
 	
 }
